Move interfaces folder to src
