+++ src/interfaces/foulInterfaces.ts
